WIP typescript
