+++ .gitignore
@@ -3,3 +3,4 @@ playsounds/*.wav
 .idea
 necrodancer-files
 .yarn
+build/
