+++ get-mp3-duration.d.ts
@@ -1,0 +1,3 @@
+declare module 'get-mp3-duration';
+// TODO get it working inside module
+// export default function getDuration(data: Buffer): number { }
